add siphash

+++ src/kernel/crypto.h
@@ -2,11 +2,7 @@
 #define CRYPTO_H
 
 
-
-
-
 uint8_t *siphash_key_gen(void);
-uint64_t siphash24(uint8_t *in, int inlen, uint8_t key[16], uint8_t out[8]);
-
+int siphash24(uint8_t *in, int inlen, uint8_t key[16], uint8_t out[8]);
 
 #endif
